requirements.txt
openai==1.58.1
pydantic==2.10.3
pydantic-settings==2.5.2
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.34.0
tenacity==9.0.0
